Generic and Specific Styling
#specific.css
This file will be used to style behaviour that should change if some text on the main page were to change.
For example, if I were to change the subtitle, the spacing of the words would change and I would need to check which screen range needs to be "adapted" to that new title.