Fixed Layouts Bug
I learned an interesting thing about jekyll layouts  today

#_layouts/light.html
This file is a layout, but it was including the home layout (copy-paste fault)!
Basically, layouts can pull layouts, which means I can have a generic layout and another layout just instances that and modifies it somehow (still not sure how to do that part though).